fix: enum passed to methods instead of Strings

+++ CultWeb/src/main/java/ch/unige/idsi/cultweb/model/RecommendationManager.java
@@ -19,7 +19,7 @@ public class RecommendationManager {
 		this.fullList.add(r3);
 	}
 	
-	public List<Recommendation> get(long id) {
+	public List<Recommendation> get(int id) {
 
 		List<Recommendation> list = new ArrayList<Recommendation>();
